+++ app/net/macolabo/sform2/models/transfer/TransferConfigMailAddress.scala
@@ -17,8 +17,8 @@ case class TransferConfigMailAddress(
                                     modified: ZonedDateTime
                                     ){
   import TransferConfigMailAddress._
-  def insert: Int = create(this)
-  def update: Int = save(this)
+  def insert: BigInt = create(this)
+  def update: BigInt = save(this)
 }
 
 object TransferConfigMailAddress extends SQLSyntaxSupport[TransferConfigMailAddress] {
@@ -45,7 +45,7 @@ object TransferConfigMailAddress extends SQLSyntaxSupport[TransferConfigMailAddr
    * @param session DB Session
    * @return TransferCongirMailAddressのリスト
    */
-  def getList(userGroup: String, transferConfigMailId: Int)(implicit session: DBSession = autoSession): List[TransferConfigMailAddress] = {
+  def getList(userGroup: String, transferConfigMailId: BigInt)(implicit session: DBSession = autoSession): List[TransferConfigMailAddress] = {
     val f = TransferConfigMailAddress.syntax("f")
     withSQL(
       select(
@@ -75,7 +75,7 @@ object TransferConfigMailAddress extends SQLSyntaxSupport[TransferConfigMailAddr
    * @param session DB Session
    * @return 作成したレコードのID
    */
-  def create(transferConfigMailAddress: TransferConfigMailAddress)(implicit session: DBSession = autoSession): Int = {
+  def create(transferConfigMailAddress: TransferConfigMailAddress)(implicit session: DBSession = autoSession): BigInt = {
     withSQL {
       val c = TransferConfigMailAddress.column
       insert.into(TransferConfigMailAddress).namedValues(
@@ -98,7 +98,7 @@ object TransferConfigMailAddress extends SQLSyntaxSupport[TransferConfigMailAddr
    * @param session DB Session
    * @return Result
    */
-  def save(transferConfigMailAddress: TransferConfigMailAddress)(implicit session: DBSession = autoSession): Int = {
+  def save(transferConfigMailAddress: TransferConfigMailAddress)(implicit session: DBSession = autoSession): BigInt = {
     withSQL {
       val c = TransferConfigMailAddress.column
       update(TransferConfigMailAddress).set(
@@ -122,7 +122,7 @@ object TransferConfigMailAddress extends SQLSyntaxSupport[TransferConfigMailAddr
    * @param session DB Session
    * @return Result
    */
-  def erase(userGroup: String, transferConfigMailAddressId: Int)(implicit session: DBSession = autoSession): Int = {
+  def erase(userGroup: String, transferConfigMailAddressId: BigInt)(implicit session: DBSession = autoSession): BigInt = {
     withSQL {
       delete.from(TransferConfigMailAddress)
         .where
